test: migrate ITReadMaskTest to use new StorageITParamRunner
Split enclosed classes into their own respective classes ITBlobReadMaskTest and ITBucketReadMaskTest
